@@ -9,25 +9,40 @@
         "email": "mirko.pagliai@gmail.com"
     }],
     "require": {
-        "php": ">=5.5.9",
-        "cakephp/cakephp": "^3.7.1",
-        "mirko-pagliai/me-cms": "^2.26.6",
-        "mirko-pagliai/php-tools": "^1.1.12"
+        "php": ">=7.2",
+        "cakephp/cakephp": "4.x-dev",
+        "cakephp/chronos": "2.x-dev as 2.0.0",
+        "ctlabvn/recaptcha": "dev-cake-4",
+        "mirko-pagliai/me-cms": "dev-cakephp4",
+        "mirko-pagliai/php-tools": "dev-php7"
     },
     "require-dev": {
-        "cakephp/cakephp-codesniffer": "^3.0",
-        "mirko-pagliai/cakephp-assets": "^1.3",
-        "phpunit/phpunit": "^5.7|^6.0"
+        "cakephp/cakephp-codesniffer": "dev-next",
+        "mirko-pagliai/cakephp-assets": "dev-cakephp4",
+        "phpunit/phpunit": "^8.0"
     },
     "repositories": [
         {
             "type": "composer",
             "url": "https://asset-packagist.org"
+        },
+        {
+            "type": "package",
+            "package": {
+                "name": "ctlabvn/recaptcha",
+                "version": "dev-cake-4",
+                "source": {
+                    "url": "https://github.com/ctlabvn/recaptcha.git",
+                    "type": "git",
+                    "reference": "cake-4"
+                }
+            }
         }
     ],
     "autoload": {
         "psr-4": {
-            "MeCmsInstagram\\": "src"
+            "MeCmsInstagram\\": "src",
+            "Recaptcha\\": "vendor/ctlabvn/recaptcha/src/"
         }
     },
     "autoload-dev": {
@@ -35,5 +50,6 @@
             "MeCmsInstagram\\Test\\": "tests",
             "App\\": "tests/test_app/TestApp"
         }
-    }
+    },
+    "minimum-stability": "dev"
 }
